+++ src/ExpressionPowerTools.Serialization/Signatures/IAnonymousTypeAdapter.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Jeremy Likness. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the repository root for license information.
 
+using System.Dynamic;
 using System.Linq.Expressions;
 using ExpressionPowerTools.Serialization.Serializers;
 
@@ -38,5 +39,12 @@ namespace ExpressionPowerTools.Serialization.Signatures
         /// <param name="lambda">The <see cref="LambdaExpression"/>.</param>
         /// <returns>The new Lambda expression.</returns>
         LambdaExpression TransformLambda(LambdaExpression lambda);
+
+        /// <summary>
+        /// Parses a type to swap (i.e. anonymous to <see cref="ExpandoObject"/>).
+        /// </summary>
+        /// <param name="memberToTransform">The type to transform.</param>
+        /// <returns>The transformed typed.</returns>
+        string MemberKeyTransformer(string memberToTransform);
     }
 }